requirements.txt
PyP100
telepot
pyyaml
Pillow
spidev
RPi.GPIO
waveshare-epaper
